+++ docs/changelogs/changelog-v0.4.0.md
@@ -5,3 +5,5 @@
 * Moved comparison operations to new function
 * Booleans can be set to comparisons.
 * [packagerelease.sh](/packagerelease.sh) now requires all tests to pass successfully in order to package for release
+* Split stsclasses.h into multiple files in an include dir
+* Started using azure pipelines for build automation
