+++ src/classes/decmethod.cpp
@@ -1,4 +1,4 @@
-#include "../core/stsclasses.h"
+#include "../include/stormscript.h"
 
 // this is only to declare methods, not to reference or use them.
