+++ src/core/set.cpp
@@ -1,4 +1,4 @@
-#include "stsclasses.h"
+#include "../include/stormscript.h"
 
 void sts::set(string command, string setto, int x) {
     for (int i = x; i<prs.size(); i++) {
