+++ src/include/core.h
@@ -1,73 +1,11 @@
 #pragma once
-#ifndef STSCLASSES_H_
-#define STSCLASSES_H_
+#ifndef CORE_H_
+#define CORE_H_
 
-#if defined(_WIN32)
-#define PLATFORM "Windows"
-#else
-#define PLATFORM "other"
-#endif
-
-#include <iostream>
-#include <string>
-#include <fstream>
-#include <vector>
-#include <cctype>
-#include <cstdlib>
-#include "../StormScriptconfig.h"
-
-using std::string;
-using std::cout;
-
-class sts;
-class stsvars;
-class stsfunc;
-
-// for the most part, it is a better idea to use functions outside of the sts class
-string striplit(string line);
-bool condition(sts *program, int *ln, std::vector<stsvars> vars);
-bool isint(string s);
-
-class stsvars{
-public:
-	int valint = 0;
-	int length;
-	std::vector<stsvars> vals;
-	string valstring = "";
-	bool val, glob;
-	string name;
-	char type;
-
-	void assignlist(sts *script, std::vector<stsvars> vars, int *line);
-};
-
-class stsclass {
-public:
-	void declare(int *y, sts *inst);
-	void decmethod(sts *inst, int *ln);
-	std::vector<stsvars> variables;
-	std::vector<stsfunc> methods;
-	string name;
-};
-
-class stsclasstype:public stsvars{
-public:
-	stsclass tpe;
-	std::vector<int> indexes;
-};
-
-class stsfunc:public stsvars{
-public:
-	int linestarted; // line funcion code starts, not line "func" keyword is used
-	bool classmethod = false;
-	stsvars value;
-	string cof;
-	string modname = "main";
-	stsclasstype cval;
-	char valtype;
-	std::vector<stsvars> args;
-};
-
+#include "includes.h"
+#include "variables.h"
+#include "functions.h"
+#include "classes.h"
 
 class sts
 {
@@ -105,7 +43,5 @@ public:
 	bool valchange(std::vector<stsvars> * pvars, std::vector<stsclasstype> *classtypes, int * ln);
 	stsvars math(int *y, std::vector<stsvars> vars);
 };
-
-
 
 #endif
