+++ src/parser/modules.cpp
@@ -1,4 +1,4 @@
-#include "../core/stsclasses.h"
+#include "../include/stormscript.h"
 
 void sts::addmodule(int *ln) {
     char **args;
