+++ src/values/man.cpp
@@ -1,4 +1,4 @@
-#include "../core/stsclasses.h"
+#include "../include/stormscript.h"
 
 /* 
 This file deals with non-keyword commands like:
